rename: compute.ts → transform.ts
